+++ src/gql/documents/mutations/account/accountAuthenticate.ts
@@ -1,8 +1,6 @@
 import { useMutation } from '@urql/vue'
 import { graphql } from '~~/gql/generated'
 
-export const useAuthenticateMutation = () => useMutation(authenticateMutation)
-
 export const authenticateMutation = graphql(`
   mutation Authenticate($password: String!, $username: String!) {
     authenticate(input: { password: $password, username: $username }) {
@@ -11,3 +9,5 @@ export const authenticateMutation = graphql(`
     }
   }
 `)
+
+export const useAuthenticateMutation = () => useMutation(authenticateMutation)