+++ src/gql/documents/mutations/account/accountDelete.ts
@@ -1,13 +1,12 @@
 import { useMutation } from '@urql/vue'
 import { graphql } from '~~/gql/generated'
 
-export const useAccountDeleteMutation = () =>
-  useMutation(
-    graphql(`
-      mutation AccountDelete($password: String!) {
-        accountDelete(input: { password: $password }) {
-          clientMutationId
-        }
-      }
-    `),
-  )
+export const accountDeleteMutation = graphql(`
+  mutation AccountDelete($password: String!) {
+    accountDelete(input: { password: $password }) {
+      clientMutationId
+    }
+  }
+`)
+
+export const useAccountDeleteMutation = () => useMutation(accountDeleteMutation)