+++ src/gql/documents/mutations/account/accountPasswordChange.ts
@@ -1,21 +1,18 @@
 import { useMutation } from '@urql/vue'
 import { graphql } from '~~/gql/generated'
 
+export const accountPasswordChangeMutation = graphql(`
+  mutation AccountPasswordChange(
+    $passwordCurrent: String!
+    $passwordNew: String!
+  ) {
+    accountPasswordChange(
+      input: { passwordCurrent: $passwordCurrent, passwordNew: $passwordNew }
+    ) {
+      clientMutationId
+    }
+  }
+`)
+
 export const useAccountPasswordChangeMutation = () =>
-  useMutation(
-    graphql(`
-      mutation AccountPasswordChange(
-        $passwordCurrent: String!
-        $passwordNew: String!
-      ) {
-        accountPasswordChange(
-          input: {
-            passwordCurrent: $passwordCurrent
-            passwordNew: $passwordNew
-          }
-        ) {
-          clientMutationId
-        }
-      }
-    `),
-  )
+  useMutation(accountPasswordChangeMutation)