+++ src/gql/documents/mutations/account/accountRegistration.ts
@@ -1,27 +1,27 @@
 import { useMutation } from '@urql/vue'
 import { graphql } from '~~/gql/generated'
 
-export const useAccountRegistrationMutation = () =>
-  useMutation(
-    graphql(`
-      mutation AccountRegistration(
-        $emailAddress: String!
-        $password: String!
-        $username: String!
-        $language: String!
-        $legalTermId: UUID!
-      ) {
-        accountRegistration(
-          input: {
-            emailAddress: $emailAddress
-            password: $password
-            username: $username
-            language: $language
-            legalTermId: $legalTermId
-          }
-        ) {
-          clientMutationId
-        }
+export const accountRegistrationMutation = graphql(`
+  mutation AccountRegistration(
+    $emailAddress: String!
+    $password: String!
+    $username: String!
+    $language: String!
+    $legalTermId: UUID!
+  ) {
+    accountRegistration(
+      input: {
+        emailAddress: $emailAddress
+        password: $password
+        username: $username
+        language: $language
+        legalTermId: $legalTermId
       }
-    `),
-  )
+    ) {
+      clientMutationId
+    }
+  }
+`)
+
+export const useAccountRegistrationMutation = () =>
+  useMutation(accountRegistrationMutation)