+++ src/gql/documents/mutations/preference/preferenceEventFormatCreate.ts
@@ -1,0 +1,19 @@
+import { useMutation } from '@urql/vue'
+import { graphql } from '~~/gql/generated'
+
+export const createPreferenceEventFormatMutation = graphql(`
+  mutation CreatePreferenceEventFormat(
+    $input: AccountPreferenceEventFormatInput!
+  ) {
+    createAccountPreferenceEventFormat(
+      input: { accountPreferenceEventFormat: $input }
+    ) {
+      accountPreferenceEventFormat {
+        ...PreferenceEventFormatItem
+      }
+    }
+  }
+`)
+
+export const useCreatePreferenceEventFormatMutation = () =>
+  useMutation(createPreferenceEventFormatMutation)