+++ src/gql/documents/queries/account/accountById.ts
@@ -2,12 +2,6 @@ import { useQuery } from '@urql/vue'
 import { graphql } from '~~/gql/generated'
 import type { AccountByIdQueryVariables } from '~~/gql/generated/graphql'
 
-export const useAccountByIdQuery = (variables: AccountByIdQueryVariables) =>
-  useQuery({
-    query: accountByIdQuery,
-    variables,
-  })
-
 export const accountByIdQuery = graphql(`
   query AccountById($id: UUID!) {
     accountById(id: $id) {
@@ -15,3 +9,9 @@ export const accountByIdQuery = graphql(`
     }
   }
 `)
+
+export const useAccountByIdQuery = (variables?: AccountByIdQueryVariables) =>
+  useQuery({
+    query: accountByIdQuery,
+    variables,
+  })