+++ src/gql/documents/queries/guest/guestsAll.ts
@@ -2,12 +2,6 @@ import { useQuery } from '@urql/vue'
 import { graphql } from '~~/gql/generated'
 import type { AllGuestsQueryVariables } from '~~/gql/generated/graphql'
 
-export const useAllGuestsQuery = (variables: AllGuestsQueryVariables) =>
-  useQuery({
-    query: allGuestsQuery,
-    variables,
-  })
-
 export const allGuestsQuery = graphql(`
   query AllGuests($after: Cursor, $eventId: UUID!, $first: Int!) {
     allGuests(after: $after, condition: { eventId: $eventId }, first: $first) {
@@ -22,3 +16,9 @@ export const allGuestsQuery = graphql(`
     }
   }
 `)
+
+export const useAllGuestsQuery = (variables?: AllGuestsQueryVariables) =>
+  useQuery({
+    query: allGuestsQuery,
+    variables,
+  })