+++ src/server/utils/constants.ts
@@ -152,8 +152,11 @@ export const GET_CSP = ({
       ...(process.env.NODE_ENV === 'development'
         ? {
             'connect-src': [
+              // TODO: use first two lines only once nuxt content fix is merged and released (https://github.com/nuxt/content/pull/3344)
               `ws://${siteUrl.hostname}:4000/ws`, // hot reload
               `wss://${siteUrl.hostname}:4000/ws`, // hot reload
+              `ws://0.0.0.0:4000/ws`, // hot reload
+              `wss://0.0.0.0:4000/ws`, // hot reload
             ],
           }
         : {}),